+++ buildSrc/src/main/kotlin/Versions.kt
@@ -5,8 +5,8 @@ object Versions {
     const val serializationPlugin = "1.8.10"
     const val serializationRuntime = "1.4.1"
     const val json = "20231013"
-    const val okhttp = "4.9.3"
+    const val okhttp = "4.12.0"
     const val evaluationCore = "2.0.0-beta.2"
     const val amplitudeAnalytics = "1.12.0"
-    const val mockito = "4.8.1"
+    const val mockk = "1.13.9"
 }