+++ src/main/kotlin/Experiment.kt
@@ -4,7 +4,7 @@ import com.amplitude.experiment.util.Logger
 import com.amplitude.experiment.util.SystemLogger
 import java.util.concurrent.Executors
 
-internal const val LIBRARY_VERSION = "1.2.4"
+internal const val LIBRARY_VERSION = "1.2.7"
 
 object Experiment {